Created a separate file for shape generation
